+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.1.0 (in development)
 ### New features
 - Tint secret doors grey for the GM to differentiate them from regular doors
+- Toggle doors between secret and normal with ctrl+click
 
 ## v1.0.1
 - When adding a door to a synchronization group adjust it's state to bring it in sync with the other doors